Scrappeur V10/requirements.txt
selenium==4.18.1
pandas==2.2.1
openpyxl==3.1.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
requests==2.31.0
fake-useragent==1.4.0
validators==0.24.0
pycryptodome==3.20.0
tldextract==5.1.1
